IS Project/frontend/requirements.txt
streamlit>=1.20.0
requests>=2.26.0
pillow